Add: `ColonyFinishDeployment` subcomponent to `ColonyHome`
