+++ src/modules/dashboard/components/ColonyHome/ColonyFinishDeployment/ColonyFinishDeployment.css.d.ts
@@ -1,0 +1,2 @@
+export const finishDeploymentBannerContainer: string;
+export const finishDeploymentBanner: string;
