+++ src/modules/dashboard/components/ColonyHome/ColonyFinishDeployment/index.ts
@@ -1,0 +1,1 @@
+export { default } from './ColonyFinishDeployment';
