+++ src/modules/dashboard/components/ColonyHome/ColonyHome.css.d.ts
@@ -7,6 +7,4 @@ export const domainsDropdownContainer: string;
 export const contentActionsPanel: string;
 export const rightAside: string;
 export const events: string;
-export const upgradeBannerContainer: string;
-export const upgradeBanner: string;
 export const loadingWrapper: string;
