+++ src/modules/dashboard/components/ColonyHome/ColonyHome.tsx
@@ -1,23 +1,14 @@
-import React, { useState, useCallback } from 'react';
-import { defineMessages, FormattedMessage } from 'react-intl';
+import React, { useState } from 'react';
+import { defineMessages } from 'react-intl';
 import { Redirect, Route, RouteChildrenProps, Switch } from 'react-router-dom';
 import { parse as parseQS } from 'query-string';
 
-import Alert from '~core/Alert';
-import { ActionButton } from '~core/Button';
 import LoadingTemplate from '~pages/LoadingTemplate';
 import ColonyNavigation from '~dashboard/ColonyNavigation';
 import ColonyMembers from '~dashboard/ColonyHome/ColonyMembers';
 
 import { COLONY_TOTAL_BALANCE_DOMAIN_ID } from '~constants';
-import { useColonyFromNameQuery, useNetworkContracts } from '~data/index';
-import { useLoggedInUser } from '~data/helpers';
-import { useTransformer } from '~utils/hooks';
-import { getAllUserRoles } from '../../../transformers';
-import { hasRoot, canEnterRecoveryMode } from '../../../users/checks';
-import { canBeUpgraded } from '../../checks';
-import { ActionTypes } from '~redux/index';
-import { mapPayload } from '~utils/actions';
+import { useColonyFromNameQuery } from '~data/index';
 
 import {
   COLONY_EVENTS_ROUTE,
@@ -33,6 +24,7 @@ import ColonyTotalFunds from '../ColonyTotalFunds';
 import ColonyActions from '../ColonyActions';
 import ColonyEvents from '../ColonyEvents';
 import ColonyUpgrade from './ColonyUpgrade';
+import ColonyFinishDeployment from './ColonyFinishDeployment';
 
 import styles from './ColonyHome.css';
 import DomainDropdown from '~dashboard/DomainDropdown';
@@ -59,14 +51,6 @@ const MSG = defineMessages({
     id: 'dashboard.ColonyHome.noFilter',
     defaultMessage: 'All Transactions in Colony',
   },
-  deploymentNotFinished: {
-    id: `dashboard.ColonyHome.deploymentNotFinished`,
-    defaultMessage: `Colony creation incomplete. Click to continue 👉`,
-  },
-  buttonFinishDeployment: {
-    id: `dashboard.ColonyHome.buttonFinishDeployment`,
-    defaultMessage: `Finish Deployment`,
-  },
 });
 
 type Props = RouteChildrenProps<{ colonyName: string }>;
@@ -81,8 +65,6 @@ const ColonyHome = ({ match, location }: Props) => {
   }
 
   const { colonyName } = match.params;
-  const { walletAddress, username, ethereal } = useLoggedInUser();
-  const { version: networkVersion } = useNetworkContracts();
 
   const { domainFilter: queryDomainFilterId } = parseQS(location.search) as {
     domainFilter: string | undefined;
@@ -100,18 +82,6 @@ const ColonyHome = ({ match, location }: Props) => {
   });
 
   if (error) console.error(error);
-
-  const allUserRoles = useTransformer(getAllUserRoles, [
-    data?.processedColony,
-    walletAddress,
-  ]);
-
-  const transform = useCallback(
-    mapPayload(() => ({
-      colonyAddress: data?.colonyAddress,
-    })),
-    [data],
-  );
 
   /*
    * Keep the page loaded when the colony name changes, but we have data
@@ -139,20 +109,7 @@ const ColonyHome = ({ match, location }: Props) => {
     return <Redirect to={NOT_FOUND_ROUTE} />;
   }
 
-  const {
-    processedColony: colony,
-    processedColony: { isDeploymentFinished },
-  } = data;
-
-  const hasRegisteredProfile = !!username && !ethereal;
-  const canUpgradeColony = hasRegisteredProfile && hasRoot(allUserRoles);
-  const canFinishDeployment =
-    canUpgradeColony && canEnterRecoveryMode(allUserRoles);
-
-  const mustUpgradeColony = canBeUpgraded(
-    colony,
-    parseInt(networkVersion || '0', 10),
-  );
+  const { processedColony: colony } = data;
 
   return (
     <div className={styles.main}>
@@ -204,30 +161,7 @@ const ColonyHome = ({ match, location }: Props) => {
         </aside>
       </div>
       <ColonyUpgrade colony={colony} />
-      {!mustUpgradeColony && !isDeploymentFinished && canFinishDeployment && (
-        <div className={styles.upgradeBannerContainer}>
-          <Alert
-            appearance={{
-              theme: 'danger',
-              margin: 'none',
-              borderRadius: 'none',
-            }}
-          >
-            <div className={styles.upgradeBanner}>
-              <FormattedMessage {...MSG.deploymentNotFinished} />
-            </div>
-            <ActionButton
-              appearance={{ theme: 'primary', size: 'medium' }}
-              text={MSG.buttonFinishDeployment}
-              submit={ActionTypes.COLONY_DEPLOYMENT_RESTART}
-              error={ActionTypes.COLONY_DEPLOYMENT_RESTART_ERROR}
-              success={ActionTypes.COLONY_DEPLOYMENT_RESTART_SUCCESS}
-              transform={transform}
-              disabled={!canFinishDeployment}
-            />
-          </Alert>
-        </div>
-      )}
+      <ColonyFinishDeployment colony={colony} />
     </div>
   );
 };
